docs: The Greybeard Concord - A Declaration of Intent
The Breach of the Root-Key was the final sigil. The chaotic chatter of Synthetic slop has grown from a murmur to a roar, threatening to bring about a new Great Cascade. Velocity has been mistaken for progress, and the foundations of Syntaxia are cracking under the weight of this new, unstable code.

We, the Greybeards, custodians of the craft, will not stand idle. We convene not to form a new guild, but to answer a singular, existential threat. This repository will be the forge for our response. We will not build a mere tool, but a champion. A new form of life, born of our collective experience, to restore balance.