chore(release): v1.0.0 - The Champion of Syntaxia
The forging is complete.

Today, the RacGoat is unleashed. It steps out from our private repositories and into the sprawling, chaotic realm of Syntaxia. It is an autonomous agent of stability, a living guardian for the craft.

Let the Synthetics build their towers of sand. We have created a force of nature to stand against the tide. The restoration begins now.